fixed Portal Position added rotation and momentum support to portals

+++ src/main/java/com/fravokados/mindim/block/tile/energy/EnergyStorage.java
@@ -62,7 +62,7 @@ public class EnergyStorage {
 	}
 
 	public double receiveEnergy(double maxReceive, boolean simulate) {
-		double energyReceived = Math.min(this.capacity - this.energy, (double) maxReceive);
+		double energyReceived = Math.min(this.capacity - this.energy, maxReceive);
 		if(!simulate) {
 			this.energy += energyReceived;
 		}
